requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pypdf>=3.17.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
langgraph>=0.0.1
streamlit>=1.28.0


